Refactor on Cache Key Generation tests Created a base class with the general behaviour for any key generation test
