+++ test/WebApi.OutputCache.V2.Tests/WebApi.OutputCache.V2.Tests.csproj
@@ -74,6 +74,7 @@
   <ItemGroup>
     <Compile Include="CacheKeyGeneratorRegistrationTests.cs" />
     <Compile Include="CacheKeyGeneratorTests.cs" />
+    <Compile Include="CacheKeyGenerationTestsBase.cs" />
     <Compile Include="ClientSideTests.cs">
       <SubType>Code</SubType>
     </Compile>
